refactor : utils
