+++ src/main/java/lotto/utils/Converter.java
@@ -1,0 +1,7 @@
+package lotto.utils;
+
+public class Converter {
+    public static int changeUnit(int totalPrice) {
+        return totalPrice / 1000;
+    }
+}
