+++ src/main/java/lotto/utils/Sorter.java
@@ -1,4 +1,11 @@
 package lotto.utils;
 
+import java.util.Collections;
+import java.util.List;
+
 public class Sorter {
+    public static List<Integer> sortNumber(List<Integer> list) {
+        Collections.sort(list);
+        return list;
+    }
 }
